// src/app/services/api/model/evento.ts
/**
 * ApiEventos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Evento { 
    idEvento?: number;
    idUsuario?: number;
    nombre?: string | null;
    descripcion?: string | null;
    fechaHora?: string;
    ubicacion?: string | null;
    capacidad?: number | null;
}